plats-du-jour/requirements.txt
playwright==1.43.0
anthropic>=0.25.0,<1.0
python-dotenv==1.0.1
requests==2.32.3
